chore: refactor announce to new style
