+++ koala/cogs/announce/models.py
@@ -1,0 +1,13 @@
+from sqlalchemy import Column, Integer, String, ForeignKey, orm
+from koala.models import mapper_registry
+
+
+@mapper_registry.mapped
+class GuildUsage:
+    __tablename__ = 'GuildUsage'
+    guild_id = Column(Integer, ForeignKey("GuildExtensions.guild_id"), primary_key=True)
+    last_message_epoch_time = Column(Integer)
+
+    def __repr__(self):
+        return "GuildUsage(%s, %s)>" % \
+               (self.guild_id, self.last_message_epoch_time)
