+++ koala/cogs/announce/utils.py
@@ -1,0 +1,5 @@
+# Constants
+ANNOUNCE_SEPARATION_DAYS = 30
+SECONDS_IN_A_DAY = 86400
+TIMEOUT_TIME = 60.0
+MAX_MESSAGE_LENGTH = 2000
